requirements.txt
aiogram==3.15.0
python-dotenv==1.0.0
aiofiles==23.2.1